tests/tier-c-stripe-monkey-patch/requirements.txt
stripe==9.9.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
